move LldpNeighbor
