+++ common/Cargo.toml
@@ -30,7 +30,6 @@ humantime.workspace = true
 iddqd.workspace = true
 ipnetwork.workspace = true
 itertools.workspace = true
-lldp_protocol.workspace = true
 macaddr.workspace = true
 omicron-ledger.workspace = true
 omicron-uuid-kinds.workspace = true
