+++ nexus/src/app/lldp.rs
@@ -11,8 +11,8 @@ use lldpd_client::types::Neighbor;
 use lldpd_client::types::PortId;
 use nexus_db_queries::context::OpContext;
 use nexus_types::external_api::networking::LldpLinkConfig;
+use nexus_types::external_api::networking::LldpNeighbor;
 use omicron_common::api::external::Error;
-use omicron_common::api::external::LldpNeighbor;
 use omicron_common::api::external::LookupResult;
 use omicron_common::api::external::Name;
 use omicron_common::api::external::UpdateResult;
