+++ nexus/types/versions/Cargo.toml
@@ -15,6 +15,7 @@ chrono.workspace = true
 daft.workspace = true
 dropshot.workspace = true
 http.workspace = true
+lldp_protocol.workspace = true
 mg-admin-client.workspace = true
 mg-api-types.workspace = true
 omicron-common.workspace = true
